// generated/src/app/api-models/case-reason-rule.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class CaseReasonRule {
  seqReasonRuleId: number;
  reasonCsCode: string;
  category1CsCode: string;
  category2CsCode: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}